+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+* sentence parts have no meaning for now
+
 ## [0.0.7] - 2022-11-27
 
 ### Changed